+++ docker-java-orchestration-core/src/main/java/com/alexecollins/docker/orchestration/DockerfileValidator.java
@@ -1,6 +1,5 @@
 package com.alexecollins.docker.orchestration;
 
-import com.alexecollins.docker.orchestration.model.Id;
 import com.github.dockerjava.api.DockerClientException;
 import com.github.dockerjava.core.GoLangFileMatch;
 import com.github.dockerjava.core.GoLangFileMatchException;
@@ -11,46 +10,46 @@ import org.apache.commons.io.FilenameUtils;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Pattern;
 
 /**
- * Dockerfile validator : validate the file format 
- * TODO add check on best practice
+ * Dockerfile validator : validate the file format
  */
-public class DockerfileValidator {
+class DockerfileValidator {
 
     // Some regexes sourced from:
     // http://stackoverflow.com/a/2821201/1216976
     // http://stackoverflow.com/a/3809435/1216976
     // http://stackoverflow.com/a/6949914/1216976
-    private final static HashMap<String, Pattern> instructionsParams = new HashMap<String, Pattern>();
+    private final static Map<String, Pattern> INSTRUCTIONS_PATTERNS = instructionsPatterns();
 
-    static {
-        instructionsParams.put("FROM", Pattern.compile("^[a-z0-9.\\/_-]+(:[a-z0-9._-]+)?$", Pattern.MULTILINE ));
-        instructionsParams.put("MAINTAINER", Pattern.compile(".+"));
-        instructionsParams.put("EXPOSE", Pattern.compile("^[0-9]+([0-9\\s]+)?$"));
-        instructionsParams.put("ENV", Pattern.compile("^[a-zA-Z_]+[a-zA-Z0-9_]* .+$"));
-        instructionsParams.put("USER", Pattern.compile("^[a-z_][a-z0-9_]{0,30}$"));
-        instructionsParams.put("RUN", Pattern.compile(".+"));
-        instructionsParams.put("CMD", Pattern.compile(".+"));
-        instructionsParams.put("ONBUILD", Pattern.compile(".+"));
-        instructionsParams.put("ENTRYPOINT", Pattern.compile(".+"));
-        instructionsParams.put("ADD", Pattern.compile("^(~?[A-z0-9\\/_.-]+|https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_\\+.~#?&\\/\\/=]*))\\s~?[A-z0-9\\/_.-]+$"));
-        instructionsParams.put("COPY", Pattern.compile("^(~?[A-z0-9\\/_.-]+|https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_\\+.~#?&\\/\\/=]*))\\s~?[A-z0-9\\/_.-]+$"));
-        instructionsParams.put("VOLUME", Pattern.compile("^~?([A-z0-9\\/_.-]+|\\[(\\s*)?(\"[A-z0-9\\/_. -]+\"(,\\s*)?)+(\\s*)?\\])$"));
-        instructionsParams.put("WORKDIR", Pattern.compile("^~?[A-z0-9\\/_.-]+$"));
+    private static Map<String, Pattern> instructionsPatterns() {
+        Pattern addPattern = Pattern.compile("^(~?[${}A-z0-9\\/_.-]+|https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_\\+.~#?&\\/\\/=]*))\\s~?[A-z0-9\\/_.-]+$");
+        Map<String, Pattern> instructionPatterns = new HashMap<String, Pattern>();
+        instructionPatterns.put("FROM", Pattern.compile("^[a-z0-9./_-]+(:[a-z0-9._-]+)?$", Pattern.MULTILINE));
+        instructionPatterns.put("MAINTAINER", Pattern.compile(".+"));
+        instructionPatterns.put("EXPOSE", Pattern.compile("^[0-9]+([0-9\\s]+)?$"));
+        instructionPatterns.put("ENV", Pattern.compile("^[a-zA-Z_]+[a-zA-Z0-9_]* .+$"));
+        instructionPatterns.put("USER", Pattern.compile("^[a-z_][a-z0-9_]{0,30}$"));
+        instructionPatterns.put("RUN", Pattern.compile(".+"));
+        instructionPatterns.put("CMD", Pattern.compile(".+"));
+        instructionPatterns.put("ONBUILD", Pattern.compile(".+"));
+        instructionPatterns.put("ENTRYPOINT", Pattern.compile(".+"));
+        instructionPatterns.put("ADD", addPattern);
+        instructionPatterns.put("COPY", addPattern);
+        instructionPatterns.put("VOLUME", Pattern.compile("^~?([A-z0-9\\/_.-]+|\\[(\\s*)?(\"[A-z0-9\\/_. -]+\"(,\\s*)?)+(\\s*)?\\])$"));
+        instructionPatterns.put("WORKDIR", Pattern.compile("^~?[A-z0-9\\/_.-]+$"));
+        return instructionPatterns;
     }
 
-
-    public static void validate(Id id, File src) throws IOException {
+    void validate(File src) throws IOException {
         Preconditions.checkArgument(src.exists(),
                 "Path %s doesn't exist", src);
         File dockerFile;
-        if(src.isDirectory()) {
+        if (src.isDirectory()) {
             Preconditions.checkState(new File(src, "Dockerfile").exists(),
                     "Dockerfile doesn't exist in " + src);
 
@@ -61,7 +60,7 @@ public class DockerfileValidator {
             dockerFile = src;
             src = dockerFile.getParentFile();
         }
-        
+
         List<String> dockerFileContent = FileUtils.readLines(dockerFile);
 
         if (dockerFileContent.size() <= 0) {
@@ -69,12 +68,11 @@ public class DockerfileValidator {
                     "Dockerfile %s is empty", dockerFile));
         }
 
-        List<String> ignores = new ArrayList<String>();
         File dockerIgnoreFile = new File(src, ".dockerignore");
         if (dockerIgnoreFile.exists()) {
             int lineNumber = 0;
             List<String> dockerIgnoreFileContent = FileUtils.readLines(dockerIgnoreFile);
-            for (String pattern: dockerIgnoreFileContent) {
+            for (String pattern : dockerIgnoreFileContent) {
                 lineNumber++;
                 pattern = pattern.trim();
                 if (pattern.isEmpty()) {
@@ -87,16 +85,11 @@ public class DockerfileValidator {
                         throw new DockerClientException(
                                 String.format("Dockerfile is excluded by pattern '%s' on line %s in .dockerignore file", pattern, lineNumber));
                     }
-                    ignores.add(pattern);
                 } catch (GoLangFileMatchException e) {
                     throw new DockerClientException(String.format("Invalid pattern '%s' on line %s in .dockerignore file", pattern, lineNumber));
                 }
             }
         }
-        List<File> filesToAdd = new ArrayList<File>();
-        filesToAdd.add(dockerFile);
-
-        Map<String, String> environmentMap = new HashMap<String, String>();
 
         int lineNumber = 0;
         boolean fromCheck = false;
@@ -106,19 +99,19 @@ public class DockerfileValidator {
             lineNumber++;
 
             if (cmd.trim().isEmpty() || cmd.startsWith("#"))
-                continue; // skip emtpy and commend lines
+                continue; // skip empty and comment lines
 
             currentLine += cmd;
-            if(cmd.endsWith("\\")) {
+            if (cmd.endsWith("\\")) {
                 continue;
             }
-            
+
             String instruction;
             String instructionParams;
-            if(currentLine.trim().contains(" ")) {
-                final String[] splitedLine = currentLine.trim().split(" ", 2);
-                instruction = splitedLine[0];
-                instructionParams = splitedLine[1];
+            if (currentLine.trim().contains(" ")) {
+                final String[] splitLine = currentLine.trim().split(" ", 2);
+                instruction = splitLine[0];
+                instructionParams = splitLine[1];
             } else {
                 instruction = currentLine.trim();
                 instructionParams = null;
@@ -139,24 +132,25 @@ public class DockerfileValidator {
             }
 
 
-            if (instructionsParams.containsKey(instruction)) {
-                if (! instructionsParams.get(instruction).matcher(instructionParams).matches())
+            if (INSTRUCTIONS_PATTERNS.containsKey(instruction)) {
+                assert instructionParams != null;
+                if (!INSTRUCTIONS_PATTERNS.get(instruction).matcher(instructionParams).matches())
                     throw new IllegalArgumentException(String.format(
                             "Wrong %s format on line [%d] of %s", currentLine, lineNumber, dockerFile));
             } else {
                 throw new IllegalArgumentException(String.format(
                         "Wrong instruction %s on line [%d] of %s", currentLine, lineNumber, dockerFile));
             }
-            
+
             //Deal with multi lines
             currentLine = "";
 
         }
-        
-        if(!Strings.isNullOrEmpty(currentLine)){
+
+        if (!Strings.isNullOrEmpty(currentLine)) {
             throw new IllegalArgumentException(String.format(
                     "Last instruction is not finish on line [%d] of %s, please remove the backslash", lineNumber, dockerFile));
-            
+
         }
 
     }